--- a/.npmignore
+++ b/.npmignore
@@ -1,0 +1,2 @@
+*.test.js
+benchmark.js