--- a/index.js
+++ b/index.js
@@ -7,7 +7,8 @@ var BOOL_PROPS = [
   'indeterminate', 'readonly', 'required', 'selected', 'willvalidate'
 ]
 
-var BOOL_PROP_PATTERN = new RegExp(' (' + BOOL_PROPS.join('|') + ')=(""|\'\')', 'ig')
+var BOOL_PROP_PATTERN = new RegExp(' (' + BOOL_PROPS.join('|') + '|onclick)=(""|\'\')', 'ig')
+var DISABLED_PATTERN = new RegExp('disabled=("true"|\'true\')', 'ig')
 
 function handleValue (value) {
   if (Array.isArray(value)) {
@@ -51,7 +52,10 @@ function stringify () {
       output += handleValue(arguments[i + 1])
     }
   }
-  output = output.replace(BOOL_PROP_PATTERN, '')
+  output = output
+    .replace(DISABLED_PATTERN, 'disabled="disabled"')
+    .replace(BOOL_PROP_PATTERN, '')
+
   // HACK: Avoid double encoding by marking encoded string
   // You cannot add properties to string literals
   // eslint-disable-next-line no-new-wrappers