--- a/index.test.js
+++ b/index.test.js
@@ -2,13 +2,14 @@
 
 const createApp = require('./app')
 const pelo = require('.')
-const test = require('ava')
+const test = require('tape')
 const minify = require('html-minifier').minify
 const minifyConfig = {
   collapseWhitespace: true
 }
 
-test('app renders as expected', t => {
+test('The app renders as expected', t => {
+  t.plan(1)
   const expected = minify(createApp().render().toString(), minifyConfig)
   pelo.replace('bel')
   const actual = minify(createApp().render().toString(), minifyConfig)