--- a/package.json
+++ b/package.json
@@ -5,14 +5,16 @@
   "main": "index.js",
   "scripts": {
     "start": "npm run build && node server.js",
-    "test": "node benchmark.js",
+    "test": "ava && node benchmark.js",
     "build": "browserify client.js > bundle.js",
     "compare": "node compare.js"
   },
   "dependencies": {},
   "devDependencies": {
+    "ava": "^0.23.0",
     "bel": "^4.6.0",
     "browserify": "^14.3.0",
+    "html-minifier": "^3.5.6",
     "nanomorph": "^5.1.3",
     "standard": "^10.0.2"
   },