… compare.js
… yarn.lock