Hello, I hope you don't mind, I massaged your code a bit and was able to squeeze out an improvement in the pelo benchmark by doing some micro-optimizations, roughly a 40% improvement.

**Before**:
```
$ yarn test                    
yarn test v1.0.2
$ node benchmark.js
# benchmark 10000 iterations
bel: 1252.718ms
pelo: 212.196ms
Done in 1.79s.
```

**After**:
```
$ yarn test                    
yarn test v1.0.2
$ ava && node benchmark.js

  1 passed
# benchmark 10000 iterations
bel: 1262.899ms
pelo: 126.894ms
Done in 2.80s.
```

I have another experimental idea I wanted to test for performance improvement, but as I read the code I became inspired to just try refactoring first.

This PR isn't perfect, I'm happy to treat it as a spike and make any changes you want to clean up the history, but there are no issues I know of.

I added a test using ava, for example, so you may prefer a different test runner, but I needed a baseline to ensure I didn't break anything.  I fixed two disparities in markup compared to bel to get a psassing test (after using html-minify to remove difference in whitespace).

1. `disabled="true"` changed to `disabled="disabled"`
2. `onclick=""` was removed (there are more cases, I'm sure, to match bel, but this got me a passing test)