@shuhei 2017-11-02
Wow, thanks for your great work! It’s a great idea to minify HTML and compare.

About test runner, I prefer to have tape like bel does. Ava looks a bit overkill to me because this project does not use babel, etc. that ava supports. Would you mind changing it?