@majgis 2017-11-04
Ok, ava was replaced with tape.  Let me know if there are any other changes you want.