@shuhei 2017-11-04
Thanks a million! Published this as `0.0.5` 🎉